Bluetooth: Move scanning update into synchronous function
With recent specification version, the setup for LE scanning or LE
advertising started to more complex than it was with Bluetooth 4.0 which
initially introduced LE support. So historically, bringing the
controller into a link layer state (being it scanning or advertising)
was composed of two HCI commands. This meant using hci_req_sync worked
nicely.

However over time, the pattern of pushing work into a workqueue and then
utilizing hci_req_sync list of 5 or more commands became the normality.
With the usage of hci_req_sync it means that all states and error need
to be predicted ahead of time. This is no longer workable. Since the
work is already pushed into a workqueue, the usage of __hci_cmd_sync is
possible and process the HCI commands one command at a time.

The idea of this patch is to push towards usage of __hci_cmd_sync and
just process the changes to scanning and advertising one command at a
time and thus being able to fail gracefully and eventually recover.

Now the problem is that the usage of hci_req_sync is widespread and
touches pretty much everything. There is also hci_req_run that first
needed to be convert to use a workqueue and also process each command
one a time.

This patch is RFC and not yet ready to be merged, it is meant as a first
starting point to see how the current code can be converted.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>