Bluetooth: Move Set Device ID command to use hci_cmd_sync_queue
This is an attempt for converting Set Device ID to use
hci_cmd_sync_queue. This command is rather simple since the
mgmt_cmd_complete is always send and the update is done unconditional,
but it shows how easy the converion from hci_req_run can be.

However create_eir function and alike would have to be moved into a
better location. This be better located in hci_br.c and nicely
abstracted.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>