Bluetooth: Rename driver .prevent_wake to .wakeup
prevent_wake logic is backward since what it is really checking is
if the device may wakeup the system or not, not that it can prevent
the system to be awaken.

Also looking on how other subsystems have the entry as power/wakeup
this also renames the force_prevent_wake to just wakeup in vhci driver.

Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>