Bluetooth: vhci: Add support extended opcode for HCI Vendor packet
Current implementation uses the Vendor packet type (0xff) with opcode
parameter. But there is no way to expand the opcode and no available bits
to use. Also it cannot be changed due to the backward compatibility
with older kernel.

THis patch adds new opcode(0x03) for HCI Vendor packet for VHCI for
extended device creation. This new opcode will not conflict with
existing legacy opcode because the legacy expects to set either bit 0 or
bit 1, but not both of bits.

It aslo requires new extra parameters for device type and flags to apply
to the VHCI device.

Signed-off-by: Tedd Ho-Jeong An <tedd.an@intel.com>