btusb: CSR chip hangs when unbound
Bluetooth Dongles with CSR chip (i.e. USB Bluetooth V4.0 Dongle by Trust)
hang when they are unbound from 'unbind' sysfs entry and can not be bound
again.

The reason is CSR chip hangs when usb configuration command with index 0
(used to unconfigure) is sent during disconnection.

To avoid this unwanted result, it is necessary not to send this command
for CSR chip when usb device is unbound. Besides, "skip_unconfigure"  sysfs
entry has been created for testing purposes with these or other devices.

Athough device is not unconfigured, it is better to avoid device hanging to
be able to operate. Even bluetooth can be previously turned off.
On the other hand, this is not important if usb device is going to be bound
again (normal behavior), i.e. with usbip.

Signed-off-by: Jose Ignacio Tornos Martinez <jtornosm@redhat.com>