Bluetooth: hci_sync: Fix attempting to suspend with unfiltered passive scan
When suspending the passive scanning _must_ have its filter_policy set
to 0x01 to use the accept list otherwise _any_ advertise report would
end up waking up the system.

In order to fix the filter_policy the code now checks hdev->suspended
when attempting to program a device address using RPA into the accept
list, print a warning since the userspace shall not mark it as wakeable
when it doesn't enable LL privacy but proceed to return success so the
proper filter_policy is used.

Link: https://bugzilla.kernel.org/show_bug.cgi?id=215768
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>