Bluetooth: btqcomsmd: Fix command timeout after setting BD address
The Bluetooth firmware seems to become unresponsive for a while after
setting the BD address. At least on recent kernel versions this often
causes timeouts for subsequent commands, e.g. the HCI reset sent by
the Bluetooth core during initialization:

	Bluetooth: hci0: Opcode 0x c03 failed: -110

Unfortunately this behavior does not seem to be documented anywhere.
Experimentation suggests that the minimum necessary delay to avoid
the problem is ~150us. However, to be sure add a sleep for > 1ms
in case it is a bit longer on other firmware versions.

Fixes: 1511cc750c3d ("Bluetooth: Introduce Qualcomm WCNSS SMD based HCI driver")
Signed-off-by: Stephan Gerhold <stephan.gerhold@kernkonzept.com>