Bluetooth: btbcm: Use strreplace() in btbcm_get_board_name()
The original code works but it's a bit iffy.  The end of loop test
should be something like "board_type[i] != '\0'" but instead it is
is "i < board_type[i]".  Fortunately, those two tests are equivalent so
long as the "board_type" is not an empty string.

It's much simpler to just call strreplace() instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>