Bluetooth: NXP: Add protocol support for NXP Bluetooth chipsets
This adds a driver based on serdev driver for the NXP BT serial
protocol based on running H:4, which can enable the built-in
Bluetooth device inside a generic NXP BT chip.

This driver has Power Save feature that will put the chip into
sleep state whenever there is no activity for 2000ms, and will
be woken up when any activity is to be initiated.

This driver enables the power save feature by default by sending
the vendor specific commands to the chip during setup.

During setup, the driver is capable of reading the bootloader
signature unique to every chip, and downloading corresponding
FW file defined in a user-space config file. The firmware file
name can be defined in DTS file as well, in which case the
user-space config file will be ignored.

Signed-off-by: Neeraj Sanjay Kale <neeraj.sanjaykale@nxp.com>