Bluetooth: NXP: Add support to download helper FW file for w8997
This adds support to download helper FW file for the legacy
NXP chipset 88w8997 for the btnxpuart driver.
This helper FW file is necessary to set the bootloader baudrate
to 3000000 after which the actual BT FW file can be downloaded.
This change helps bring the FW download time from around 10 sec
to less than 2 sec for 88w8997 chip.
For newer chipsets, both V1 and V3 bootloader, driver sends the
cmd5 and cmd7 to the chip bootloader, and does not need a helper
FW file.
This also adds a call to ps_wakeup() before closing the serdev
device, to de-assert UART break. On lpuart platforms if the UART
port is closed while break is asserted, tx_empty is never set, and
waits for 30 second timeout before lpuart32_shutdown is called.
This resulted in 'hciconfig hci0 down' command to take 30 seconds
to execute.

Signed-off-by: Neeraj Sanjay Kale <neeraj.sanjaykale@nxp.com>