Bluetooth: Fix atomicity violation in {conn,adv}_{min,max}_interval_set
In {conn,adv}_min_interval_set():
	if (val < ... || val > ... || val > hdev->le_{conn,adv}_max_interval)
		return -EINVAL;
	hci_dev_lock(hdev);
	hdev->le_{conn,adv}_min_interval = val;
	hci_dev_unlock(hdev);

In {conn,adv}_max_interval_set():
	if (val < ... || val > ... || val < hdev->le_{conn,adv}_min_interval)
		return -EINVAL;
	hci_dev_lock(hdev);
	hdev->le_{conn,adv}_max_interval
	hci_dev_unlock(hdev);

The atomicity violation occurs due to concurrent execution of set_min and
set_max funcs which may lead to inconsistent reads and writes of the min
value and the max value. The checks for value validity are ineffective as
the min/max values could change immediately after being checked, raising
the risk of the min value being greater than the max value and causing
invalid settings.

This possible bug is found by an experimental static analysis tool
developed by our team, BassCheck[1]. This tool analyzes the locking APIs
to extract function pairs that can be concurrently executed, and then
analyzes the instructions in the paired functions to identify possible
concurrency bugs including data races and atomicity violations. The above
possible bug is reported when our tool analyzes the source code of
Linux 5.17.

To resolve this issue, it is suggested to encompass the validity checks
within the locked sections in both set_min and set_max funcs. The
modification ensures that the validation of 'val' against the
current min/max values is atomic, thus maintaining the integrity of the
settings. With this patch applied, our tool no longer reports the bug,
with the kernel configuration allyesconfig for x86_64. Due to the lack of
associated hardware, we cannot test the patch in runtime testing, and just
verify it according to the code logic.

[1] https://sites.google.com/view/basscheck/

Fixes: 3a5c82b78fd28 ("Bluetooth: Move LE debugfs file creation into ...")
Cc: stable@vger.kernel.org
Reported-by: BassCheck <bass@buaa.edu.cn>
Signed-off-by: Gui-Dong Han <2045gemini@gmail.com>