Bluetooth: ISO: add stream synchronization mechanism
ISO streams in the same group should be synchronized, this is required
e.g. by stereo audio playback in TWS headsets.  Currently the only
synchronization mechanism we use is that the user has to supply data for
each stream at the right time and in right order, and then we send them
to the controller ASAP without explicitly trying to keep streams in sync
and hope for the best.

However, it is observed e.g. with Intel AX210 that rather often a packet
for one of the streams does not get sent but is left in the queue and
postponed, even if submitted to controller queue at the same time (as
far as able). After this, that ISO stream latency is persistently one
packet larger relative to the other streams in the group, and e.g. TWS
stereo audio is ruined.  User sending packets to the ISO sockets cannot
know this occurred, nor can it easily correct for it.

The issue can e.g. be triggered by bad transmission (e.g. cover one of
the TWS earbuds with hand, or move the device further away), which
appears to make a random change to the latency of its stream.

Fix the problem by adding an explicit stream synchronization mechanism
to ISO sockets, and handling stream synchronization in packet
scheduling.

The user may tag each packet sent to ISO socket via CMSG with a
monotonically increasing sequence number.  The kernel shall ensure that
SDU timings within ISO group are such that packets with same CMSG
sequence number are aligned.

The implementation here does this by delaying sending some packets until
the streams are aligned.  It will also wait for controller queue to
flush sufficiently, if it appears that streams are out of sync on the
controller side.  See the code for explanation.

This patch fixes the observed TWS desynchronization issue (currently not
completely, see below), if the sound server is modified to add the
packet seq,
https://gitlab.freedesktop.org/pvir/pipewire/-/commits/iso-seqnum

I tried various other things to try to fix this (FIFO scheduling,
different number of packets submitted at the same time, etc.), without
much success, and this is the first one to give results.  Requiring
conn->sent alignment before sending more stuff to the controller seemed
the important step here.

There are a few things to think about, hence RFC:

- Any better ideas?

  The spec seems vague here (if we can't set SDU timestamps), so some
  approach that's not based on guessing what the controller internally
  does could be better.

- The conn->sent alignment check probably should be somehow refined.  It
  seems to be enough with low transport latency, but with larger
  transport latencies one can still get streams out of sync,
  characterized by the controller not reporting Number of Completed
  packets for all streams in one go, but sending completed packet events
  for some of the streams only a bit later.

- This sort of scheduling thing that depends on controller queue or
  clock IMO needs to be done in kernel, user should just tell the kernel
  how stuff is supposed to be scheduled.

- However, not clear if packet sequence number is the best way to for
  this sort of thing, or if user should specify some sort of timestamps
  or other information instead.

  E.g. we might change this to use SDU timestamps (if they can be used),
  but user would communicate the scheduling info still the same way.

- Ideally we probably should use SDU timestamps to explicitly tell the
  controller which packets go to same interval, but at least when last
  tried AX210 didn't support HCI_LE_Read_ISO_TX_Sync, and it appears
  clock synchronization is required which cannot(?) be obtained without
  it.

- This patch can make latency to grow on transmission glitch, if it
  needs to wait for controller queue to drain so that streams can be
  aligned again. For proper flow control especially in audio use cases,
  but also more generally, would be better to have for user a way to
  know the number of packets in queue, in socket buf, and in controller,
  and the max.  TIOCOUTQ is similar, but does not give packet counts nor
  controller queue so is insufficient.

Link: https://github.com/bluez/bluez/issues/515
Signed-off-by: Pauli Virtanen <pav@iki.fi>