Bluetooth: btusb: add sysfs attribute to control USB alt setting
When a Bluetooth raw socket is open, the HCI event related to SCO
connection changes are not dispatched to the hci_event module, and
the underlying Bluetooth controller's USB Interface 1 will not be
updated accordingly.

This patch adds `isoc_alt` sysfs attribute, allowing user space
to update the alternate setting of the USB interface alternate
setting as needed.

BUG=b:382526205
TEST=read and write /sys/class/bluetooth/hci0/isoc_alt

Series-to: linux-bluetooth@vger.kernel.org
Series-to: luiz.dentz@gmail.com
Series-cc: chromeos-bluetooth-upstreaming@chromium.org

Commit-notes:
This commit has been tested on a chromebook with AX211.
END

Change-Id: Ifc708cc471a8834b344c26fce1ce2fe3e5992cad
Signed-off-by: Ying Hsu <yinghsu@chromium.org>