ARM: dts: at91-sama5d27_wlsom1: update bluetooth chip description
The SAMA5D27 WLSOM1 plaftorm embeds a wilc3000 chip exposing 802.11b/g/n
and bluetooth 5. The current description for this chip does not expose
the bluetooth part: it only configures the relevant flexcom peripheral
to set it as uart, and relies on user to trigger the corresponding hci
device creation in kernel through hci_attach. Now that a bluetooth
driver is available and handles the device creation, replace the bare
uart description with a proper bluetooth node.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>