Bluetooth: btnxpuart: Fix baudrate reset on remove
Trying to reset the baudrate from device remove callback does not
work if HCI dev is down and not running, causing the next device
probing to fail, as host/controller baudrates are then out-of-sync.

To prevent this issue, we ensure the controller's baudrate is always
reset to its initial value in the shutdown callback, during the hdev
close procedure. That guarantees subsequent open procedure to succeed,
regardless of any device removal/binding in between. This also means
we want to reconfigure the baudrate in the next hdev open procedure
via the setup() callback (HCI_QUIRK_NON_PERSISTENT_SETUP required).

In case of removal when the hdev is up and running, we have to call
the shutdown procedure explicitly before unregistering the hdev.

Signed-off-by: Loic Poulain <loic.poulain@linaro.org>